update project, fix create rask bug

+++ HITSBlazor/Components/TaskComponent/TaskComponent.razor.cs
@@ -17,7 +17,10 @@ namespace HITSBlazor.Components.TaskComponent
         private ModalService ModalService { get; set; } = null!;
 
         [Parameter]
-        public required HITSTask Task { get; set; }
+        public bool IsLoading { get; set; } = false;
+
+        [Parameter]
+        public HITSTask? Task { get; set; }
 
         [Parameter]
         public EventCallback<HITSTask> TaskChanged { get; set; }
@@ -40,15 +43,22 @@ namespace HITSBlazor.Components.TaskComponent
         [Parameter]
         public int TaskWidth { get; set; } = 100;
 
-        public static (string Hint, string Color) GetHintAndColor(HITSTaskStatus status) => status switch
+        public (string Hint, string Color) GetHintAndColor()
         {
-            HITSTaskStatus.InBackLog => ("Приоритетность задачи", "bg-primary"),
-            HITSTaskStatus.Done => ("Задача завершена", "bg-secondary"),
-            _ => ("Задача в активном спринте", "bg-warning")
-        };
+            if (IsLoading || Task is null) return ("Задачи загружаются", "bg-secondary");
+
+            return Task.Status switch
+            {
+                HITSTaskStatus.InBackLog => ("Приоритетность задачи", "bg-primary"),
+                HITSTaskStatus.Done => ("Задача завершена", "bg-secondary"),
+                _ => ("Задача в активном спринте", "bg-warning")
+            };
+        }
 
         private async Task ClickTask(Func<Task> toggleMethod)
         {
+            if (IsLoading) return;
+
             if (OnClickTask.HasDelegate)
                 await OnClickTask.InvokeAsync();
             else
@@ -58,15 +68,20 @@ namespace HITSBlazor.Components.TaskComponent
 
         private void ShowTaskModal()
         {
-            if (ShowTaskModalAllowed)
-                ModalService.ShowTaskModal(Task);
+            if (ShowTaskModalAllowed && Task is not null)
+                ModalService.ShowTaskModal(Task.ProjectId, Task);
         }
 
-        private void ShowConfirmModal() => ModalService.ShowConfirmModal(
-            $"Вы действительно хотите удалить \"{Task.Name}\"?",
-            () => ProjectService.DeleteTaskAsync(Task),
-            confirmButtonVariant: ButtonVariant.Danger,
-            confirmButtonText: "Удалить"
-        );
+        private void ShowConfirmModal()
+        {
+            if (Task is null) return;
+
+            ModalService.ShowConfirmModal(
+                $"Вы действительно хотите удалить \"{Task.Name}\"?",
+                () => ProjectService.DeleteTaskAsync(Task),
+                confirmButtonVariant: ButtonVariant.Danger,
+                confirmButtonText: "Удалить"
+            );
+        }
     }
 }
